+++ src/engine_api.rs
@@ -1357,9 +1357,9 @@ impl Engine {
 
         // Trims the JSON string and add a '#' in front
         let json_text = json.trim_start();
-        let scripts = if json_text.starts_with(Token::MapStart.keyword_syntax()) {
+        let scripts = if json_text.starts_with(Token::MapStart.literal_syntax()) {
             [json_text, ""]
-        } else if json_text.starts_with(Token::LeftBrace.keyword_syntax()) {
+        } else if json_text.starts_with(Token::LeftBrace.literal_syntax()) {
             ["#", json_text]
         } else {
             return Err(crate::ParseErrorType::MissingToken(